+++ config/me_cms.php
@@ -61,11 +61,6 @@ return ['MeCms' => [
         //  recipient for the email sent by users
         'webmaster' => 'email@example.com',
     ],
-    //KCFinder
-    'kcfinder' => [
-        //Here you can rewrite the configuration for KCFinder.
-        //See http://kcfinder.sunhater.com/install#_types
-    ],
     //Main
     'main' => [
         //Date formats.