+++ src/Controller/Admin/PagesController.php
@@ -65,20 +65,6 @@ class PagesController extends AppController
     }
 
     /**
-     * Initialization hook method
-     * @return void
-     */
-    public function initialize(): void
-    {
-        parent::initialize();
-
-        //Loads KcFinderComponent
-        if ($this->getRequest()->isAction(['add', 'edit'])) {
-            $this->loadComponent('MeCms.KcFinder');
-        }
-    }
-
-    /**
      * Check if the provided user is authorized for the request
      * @param array|\ArrayAccess|null $user The user to check the authorization
      *  of. If empty the user in the session will be used